Lib convention: name files by execution context
Codifies a single naming convention for files inside lib/<domain>/:
- index.ts — universal modules
- server.ts — server-only modules
- client.ts — "use client" modules
- action.ts — "use server" actions

AGENTS.md gets a new "lib/<domain>/ files" subsection plus two
exceptions (folders that group multiple modules of the same execution
context, like lib/markdown/ and lib/agent/tools/; and flat single-file
modules at lib/ root). Also adds a guardrail: don't use "client" in a
filename to mean an HTTP/SDK client wrapper — collides with the runtime
meaning.

Migrations:
- lib/agent/{create-agent,context,prompt}.ts → lib/agent/server.ts
  (single server module with section headers; previously three tightly
  coupled server-only files)
- lib/agent/ui/catalog.ts → lib/agent/index.ts (universal — used by
  both server.ts and the "use client" registry component). ui/ removed.
- lib/auth/auth.ts → merged into lib/auth/server.ts (all server-only
  better-auth config; one file is enough).
- lib/shopify/client.ts → lib/shopify/fetch.ts (the "client" name
  collided with the runtime convention; descriptive rename).
- lib/agent/tools/*.ts (10 files): from "../context" → from "../server".

Importers in app/api/chat, app/api/auth, account layouts, nav,
agent/registry, and shopify operations updated accordingly. No public
API or behavior changes.
