+++ apps/template/app/api/auth/[...all]/route.ts
@@ -1,4 +1,5 @@
-import { auth } from "@/lib/auth/auth";
 import { toNextJsHandler } from "better-auth/next-js";
+
+import { auth } from "@/lib/auth/server";
 
 export const { GET, POST } = toNextJsHandler(auth);
