+++ apps/template/components/layout/nav/index.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { Suspense } from "react";
 
-import { isAuthEnabled } from "@/lib/auth/auth";
+import { isAuthEnabled } from "@/lib/auth/server";
 
 import { NavAccount, NavAccountFallback } from "./account";
 import { CartIcon, CartIconFallback } from "./cart";
