+++ apps/template/lib/agent/tools/add-cart-note.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 
 import { updateCartNote } from "@/lib/shopify/operations/cart";
 
-import { getAgentContext } from "../context";
+import { getAgentContext } from "../server";
 
 export function addCartNoteTool() {
   return tool({
