+++ apps/template/lib/agent/tools/add-to-cart.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 
 import { addToCart } from "@/lib/shopify/operations/cart";
 
-import { getAgentContext } from "../context";
+import { getAgentContext } from "../server";
 
 export function addToCartTool() {
   return tool({
