+++ apps/template/lib/agent/tools/get-cart.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 
 import { getCart } from "@/lib/shopify/operations/cart";
 
-import { getAgentContext } from "../context";
+import { getAgentContext } from "../server";
 
 export function getCartTool() {
   return tool({
