+++ apps/template/lib/shopify/operations/menu.ts
@@ -2,7 +2,7 @@ import { cacheLife, cacheTag } from "next/cache";
 
 import { defaultLocale } from "@/lib/i18n";
 
-import { shopifyFetch } from "../client";
+import { shopifyFetch } from "../fetch";
 import type { Menu, MenuItem, MenuItemType } from "../types/menu";
 import { transformShopifyMenuItemUrl } from "../utils";
